refs #777   *  필터 관련 체크업 루틴이 잘못 실행되는 오류 수저

+++ interface/blog/checkup.php
@@ -393,7 +393,8 @@ if($currentVersion != TEXTCUBE_VERSION && in_array(POD::dbms(),array('MySQL','My
 			showCheckupMessage(false);
 	}
 	/* From Textcube 1.7.6 */
-	if (!strpos(POD::queryCell("DESC {$database['prefix']}Filters type", 'Type'),'whiteurl')) {
+	if (!POD::queryExistence("DESC {$database['prefix']}Filters filtertype") &&
+		!strpos(POD::queryCell("DESC {$database['prefix']}Filters type", 'Type'),'whiteurl')) {
 		$changed = true;
 		echo '<li>', _text('필터 테이블에 예외 목록을 추가하기 위하여 필드 속성을 변경합니다.'), ': ';
 		if (POD::execute("ALTER TABLE {$database['prefix']}Filters CHANGE type type ENUM('content','ip','name','url','whiteurl') NOT NULL DEFAULT 'content'"))
